Add hooks for section container, section content and hidden input
