+++ packages/x-date-pickers/src/internals/hooks/useField/useFieldV7TextField.ts
@@ -2,8 +2,7 @@ import * as React from 'react';
 import useForkRef from '@mui/utils/useForkRef';
 import useEventCallback from '@mui/utils/useEventCallback';
 import useEnhancedEffect from '@mui/utils/useEnhancedEffect';
-import useId from '@mui/utils/useId';
-import { getSectionValueNow, getSectionValueText, parseSelectedSections } from './useField.utils';
+import { parseSelectedSections } from './useField.utils';
 import {
   UseFieldDOMGetters,
   UseFieldParameters,
@@ -14,14 +13,15 @@ import { getActiveElement } from '../../utils/utils';
 import { FieldSectionType } from '../../../models';
 import { useSplitFieldProps } from '../../../hooks';
 import { PickersSectionElement, PickersSectionListRef } from '../../../PickersSectionList';
-import { usePickerTranslations } from '../../../hooks/usePickerTranslations';
-import { useUtils } from '../useUtils';
 import { useFieldCharacterEditing } from './useFieldCharacterEditing';
 import { useFieldState } from './useFieldState';
 import { useFieldInternalPropsWithDefaults } from './useFieldInternalPropsWithDefaults';
 import { PickerValidValue } from '../../models';
 import { syncSelectionToDOM } from './syncSelectionToDOM';
 import { useFieldRootProps } from './useFieldRootProps';
+import { useFieldHiddenInputProps } from './useFieldHiddenInputProps';
+import { useFieldSectionContainerProps } from './useFieldSectionContainerProps';
+import { useFieldSectionContentProps } from './useFieldSectionContentProps';
 
 export const useFieldV7TextField = <
   TValue extends PickerValidValue,
@@ -31,19 +31,11 @@ export const useFieldV7TextField = <
 >(
   parameters: UseFieldParameters<TValue, true, TError, TValidationProps, TProps>,
 ): UseFieldReturnValue<true, TProps> => {
-  const translations = usePickerTranslations();
-  const utils = useUtils();
-  const id = useId();
-
   const {
     props,
     manager,
     skipContextFieldRefAssignment,
-    manager: {
-      valueType,
-      internal_fieldValueManager: fieldValueManager,
-      internal_useOpenPickerButtonAriaLabel: useOpenPickerButtonAriaLabel,
-    },
+    manager: { valueType, internal_useOpenPickerButtonAriaLabel: useOpenPickerButtonAriaLabel },
   } = parameters;
 
   const { internalProps, forwardedProps } = useSplitFieldProps(props, valueType);
@@ -97,17 +89,12 @@ export const useFieldV7TextField = <
     error,
     parsedSelectedSections,
     sectionOrder,
-    sectionsValueBoundaries,
     state,
     value,
 
     // Methods to update the states
     clearValue,
-    clearActiveSection,
-    setCharacterQuery,
     setSelectedSections,
-    updateSectionValue,
-    updateValueFromValueStr,
   } = stateResponse;
 
   const applyCharacterEditing = useFieldCharacterEditing({ stateResponse });
@@ -132,34 +119,6 @@ export const useFieldV7TextField = <
     sectionListRef.current.getSectionContent(newParsedSelectedSections).focus();
   }
 
-  /**
-   * If a section content has been updated with a value we don't want to keep,
-   * Then we need to imperatively revert it (we can't let React do it because the value did not change in his internal representation).
-   */
-  const revertDOMSectionChange = useEventCallback((sectionIndex: number) => {
-    if (!sectionListRef.current) {
-      return;
-    }
-
-    const section = state.sections[sectionIndex];
-
-    sectionListRef.current.getSectionContent(sectionIndex).innerHTML =
-      section.value || section.placeholder;
-    syncSelectionToDOM({ focused, domGetters, stateResponse });
-  });
-
-  const getInputContainerClickHandler = useEventCallback(
-    (sectionIndex: number) => (event: React.MouseEvent<HTMLDivElement>) => {
-      // The click event on the clear button would propagate to the input, trigger this handler and result in a wrong section selection.
-      // We avoid this by checking if the call to this function is actually intended, or a side effect.
-      if (event.isDefaultPrevented()) {
-        return;
-      }
-
-      setSelectedSections(sectionIndex);
-    },
-  );
-
   const rootProps = useFieldRootProps({
     manager,
     internalPropsWithDefaults,
@@ -168,6 +127,16 @@ export const useFieldV7TextField = <
     focused,
     setFocused,
     domGetters,
+  });
+  const hiddenInputProps = useFieldHiddenInputProps({ manager, stateResponse });
+  const createSectionContainerProps = useFieldSectionContainerProps({ stateResponse });
+  const createSectionContentProps = useFieldSectionContentProps({
+    manager,
+    stateResponse,
+    applyCharacterEditing,
+    internalPropsWithDefaults,
+    domGetters,
+    focused,
   });
 
   const handleRootKeyDown = useEventCallback((event: React.KeyboardEvent<HTMLDivElement>) => {
@@ -200,96 +169,6 @@ export const useFieldV7TextField = <
     rootProps.onInput(event);
   });
 
-  const handleInputContentMouseUp = useEventCallback((event: React.MouseEvent) => {
-    // Without this, the browser will remove the selected when clicking inside an already-selected section.
-    event.preventDefault();
-  });
-
-  const getInputContentFocusHandler = useEventCallback((sectionIndex: number) => () => {
-    setSelectedSections(sectionIndex);
-  });
-
-  const handleInputContentPaste = useEventCallback(
-    (event: React.ClipboardEvent<HTMLSpanElement>) => {
-      // prevent default to avoid the input `onInput` handler being called
-      event.preventDefault();
-
-      if (readOnly || disabled || typeof parsedSelectedSections !== 'number') {
-        return;
-      }
-
-      const activeSection = state.sections[parsedSelectedSections];
-      const pastedValue = event.clipboardData.getData('text');
-      const lettersOnly = /^[a-zA-Z]+$/.test(pastedValue);
-      const digitsOnly = /^[0-9]+$/.test(pastedValue);
-      const digitsAndLetterOnly = /^(([a-zA-Z]+)|)([0-9]+)(([a-zA-Z]+)|)$/.test(pastedValue);
-      const isValidPastedValue =
-        (activeSection.contentType === 'letter' && lettersOnly) ||
-        (activeSection.contentType === 'digit' && digitsOnly) ||
-        (activeSection.contentType === 'digit-with-letter' && digitsAndLetterOnly);
-
-      if (isValidPastedValue) {
-        setCharacterQuery(null);
-        updateSectionValue({
-          section: activeSection,
-          newSectionValue: pastedValue,
-          shouldGoToNextSection: true,
-        });
-      }
-      // If the pasted value corresponds to a single section, but not the expected type, we skip the modification
-      else if (!lettersOnly && !digitsOnly) {
-        setCharacterQuery(null);
-        updateValueFromValueStr(pastedValue);
-      }
-    },
-  );
-
-  const handleInputContentDragOver = useEventCallback((event: React.DragEvent<HTMLSpanElement>) => {
-    event.preventDefault();
-    event.dataTransfer.dropEffect = 'none';
-  });
-
-  const handleInputContentInput = useEventCallback((event: React.FormEvent<HTMLSpanElement>) => {
-    if (!sectionListRef.current) {
-      return;
-    }
-
-    const target = event.target as HTMLSpanElement;
-    const keyPressed = target.textContent ?? '';
-    const sectionIndex = sectionListRef.current.getSectionIndexFromDOMElement(target)!;
-    const section = state.sections[sectionIndex];
-
-    if (readOnly || !sectionListRef.current) {
-      revertDOMSectionChange(sectionIndex);
-      return;
-    }
-
-    if (keyPressed.length === 0) {
-      if (section.value === '') {
-        revertDOMSectionChange(sectionIndex);
-        return;
-      }
-
-      const inputType = (event.nativeEvent as InputEvent).inputType;
-      if (inputType === 'insertParagraph' || inputType === 'insertLineBreak') {
-        revertDOMSectionChange(sectionIndex);
-        return;
-      }
-
-      revertDOMSectionChange(sectionIndex);
-      clearActiveSection();
-      return;
-    }
-
-    applyCharacterEditing({
-      keyPressed,
-      sectionIndex,
-    });
-
-    // The DOM value needs to remain the one React is expecting.
-    revertDOMSectionChange(sectionIndex);
-  });
-
   const handleClear = useEventCallback((event: React.MouseEvent, ...args) => {
     event.preventDefault();
     onClear?.(event, ...(args as []));
@@ -303,88 +182,18 @@ export const useFieldV7TextField = <
     }
   });
 
-  const isContainerEditable = parsedSelectedSections === 'all';
   const elements = React.useMemo<PickersSectionElement[]>(() => {
-    return state.sections.map((section, index) => {
-      const isEditable = !isContainerEditable && !disabled && !readOnly;
-      const sectionBoundaries = sectionsValueBoundaries[section.type]({
-        currentDate: fieldValueManager.getDateFromSection(value, section),
-        contentType: section.contentType,
-        format: section.format,
-      });
-
-      return {
-        container: {
-          'data-sectionindex': index,
-          onClick: getInputContainerClickHandler(index),
-        } as React.HTMLAttributes<HTMLSpanElement>,
-        content: {
-          tabIndex: isContainerEditable || index > 0 ? -1 : 0,
-          contentEditable: !isContainerEditable && !disabled && !readOnly,
-          role: 'spinbutton',
-          id: `${id}-${section.type}`,
-          'aria-labelledby': `${id}-${section.type}`,
-          'aria-readonly': readOnly,
-          'aria-valuenow': getSectionValueNow(section, utils),
-          'aria-valuemin': sectionBoundaries.minimum,
-          'aria-valuemax': sectionBoundaries.maximum,
-          'aria-valuetext': section.value
-            ? getSectionValueText(section, utils)
-            : translations.empty,
-          'aria-label': translations[section.type],
-          'aria-disabled': disabled,
-          spellCheck: isEditable ? false : undefined,
-          autoCapitalize: isEditable ? 'off' : undefined,
-          autoCorrect: isEditable ? 'off' : undefined,
-          [parseInt(React.version, 10) >= 17 ? 'enterKeyHint' : 'enterkeyhint']: isEditable
-            ? 'next'
-            : undefined,
-          children: section.value || section.placeholder,
-          onInput: handleInputContentInput,
-          onPaste: handleInputContentPaste,
-          onFocus: getInputContentFocusHandler(index),
-          onDragOver: handleInputContentDragOver,
-          onMouseUp: handleInputContentMouseUp,
-          inputMode: section.contentType === 'letter' ? 'text' : 'numeric',
-        },
-        before: {
-          children: section.startSeparator,
-        },
-        after: {
-          children: section.endSeparator,
-        },
-      };
-    });
-  }, [
-    disabled,
-    fieldValueManager,
-    getInputContainerClickHandler,
-    getInputContentFocusHandler,
-    handleInputContentDragOver,
-    handleInputContentInput,
-    handleInputContentMouseUp,
-    handleInputContentPaste,
-    id,
-    isContainerEditable,
-    readOnly,
-    sectionsValueBoundaries,
-    state.sections,
-    translations,
-    utils,
-    value,
-  ]);
-
-  const handleValueStrChange = useEventCallback((event: React.ChangeEvent<HTMLInputElement>) => {
-    updateValueFromValueStr(event.target.value);
-  });
-
-  const valueStr = React.useMemo(
-    () =>
-      areAllSectionsEmpty
-        ? ''
-        : fieldValueManager.getV7HiddenInputValueFromSections(state.sections),
-    [areAllSectionsEmpty, state.sections, fieldValueManager],
-  );
+    return state.sections.map((section, sectionIndex) => ({
+      container: createSectionContainerProps(sectionIndex),
+      content: createSectionContentProps(section, sectionIndex),
+      before: {
+        children: section.startSeparator,
+      },
+      after: {
+        children: section.endSeparator,
+      },
+    }));
+  }, [state.sections, createSectionContainerProps, createSectionContentProps]);
 
   React.useEffect(() => {
     if (sectionListRef.current == null) {
@@ -458,6 +267,9 @@ export const useFieldV7TextField = <
     onKeyDown: handleRootKeyDown,
     onClear: handleClear,
 
+    // Hidden input props
+    ...hiddenInputProps,
+
     error,
     clearable: Boolean(clearable && !areAllSectionsEmpty && !readOnly && !disabled),
     focused: focusedProp ?? focused,
@@ -465,8 +277,6 @@ export const useFieldV7TextField = <
     // Additional
     enableAccessibleFieldDOMStructure: true,
     elements,
-    value: valueStr,
-    onChange: handleValueStrChange,
     areAllSectionsEmpty,
     disabled,
     readOnly,
